# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Greenways API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PasswordInput {
  password: string;
}

export interface PasswordResult {
  valid: boolean;
}

export interface FaceScanInput {
  /** Base64-encoded face image */
  imageData: string;
}

export interface FaceScanResult {
  sessionId: string;
}

export type AuthStatusStatus = typeof AuthStatusStatus[keyof typeof AuthStatusStatus];


export const AuthStatusStatus = {
  pending: 'pending',
  approved: 'approved',
  denied: 'denied',
} as const;

export interface AuthStatus {
  sessionId: string;
  status: AuthStatusStatus;
  /** @nullable */
  sessionToken?: string | null;
}

export interface VideoInput {
  title: string;
  description?: string;
  sessionToken: string;
}

export interface Video {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  filename: string;
  fileUrl?: string;
  createdAt: string;
}

export interface VideoStats {
  totalVideos: number;
  recentUploads: number;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminLoginResult {
  valid: boolean;
  /** @nullable */
  adminToken?: string | null;
}

export type ApprovalStatus = typeof ApprovalStatus[keyof typeof ApprovalStatus];


export const ApprovalStatus = {
  pending: 'pending',
  approved: 'approved',
  denied: 'denied',
} as const;

export interface Approval {
  id: number;
  sessionId: string;
  /** @nullable */
  faceImageData?: string | null;
  status: ApprovalStatus;
  createdAt: string;
}

export type ListPendingApprovalsParams = {
adminToken?: string;
};