# lib/api-zod/src/generated/types/approval.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Greenways API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApprovalStatus } from './approvalStatus';

export interface Approval {
  id: number;
  sessionId: string;
  /** @nullable */
  faceImageData?: string | null;
  status: ApprovalStatus;
  createdAt: string;
}