# lib/api-zod/src/generated/types/authStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Greenways API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthStatusStatus } from './authStatusStatus';

export interface AuthStatus {
  sessionId: string;
  status: AuthStatusStatus;
  /** @nullable */
  sessionToken?: string | null;
}