# lib/api-zod/src/generated/types/faceScanInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Greenways API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FaceScanInput {
  /** Base64-encoded face image */
  imageData: string;
}