# lib/api-zod/src/generated/types/video.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Greenways API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Video {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  filename: string;
  fileUrl?: string;
  createdAt: string;
}