# lib/api-zod/src/generated/types/videoStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Greenways API specification
 * OpenAPI spec version: 0.1.0
 */

export interface VideoStats {
  totalVideos: number;
  recentUploads: number;
}